requirements/requirements-auth.txt
# Authentication System Requirements
# JWT and Cryptography
PyJWT>=2.8.0
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0

# Password Hashing
bcrypt>=4.1.2
argon2-cffi>=23.1.0

# Email Validation
email-validator>=2.1.0

# MFA/2FA Support
pyotp>=2.9.0
qrcode>=7.4.2
Pillow>=10.2.0

# OAuth2 Support
authlib>=1.3.0

# Redis for session/token management
redis>=5.0.1

# Database
SQLAlchemy>=2.0.25
psycopg2-binary>=2.9.9

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Rate Limiting
slowapi>=0.1.9

# HTTP Client for OAuth
httpx>=0.26.0

# Environment Management
python-dotenv>=1.0.0
